Extracted stringReader
