+++ src/test/jsonTokenizer.test.ts
@@ -1,40 +1,8 @@
 import * as fs from 'fs';
 import * as path from 'path';
 import {TokenHandler} from 'tokenizer-dsl';
-import {jsonTokenizer, stringReader, Type} from '../main/jsonTokenizer';
+import {jsonTokenizer, Type} from '../main/jsonTokenizer';
 import {ParserContext} from '../main/parser-types';
-
-describe('stringReader', () => {
-
-  let context: ParserContext;
-
-  beforeEach(() => {
-    context = {
-      stack: [],
-      cursor: 0,
-      arrayMode: false,
-      objectKey: '',
-      input: '',
-      parseBigInt: BigInt,
-    };
-  });
-
-  test('reads string', () => {
-    expect(stringReader('"aaa"', 0, context)).toBe(5);
-  });
-
-  test('reads string with escaped quote chars', () => {
-    expect(stringReader('"aaa\\"bbb\\""', 0, context)).toBe(12);
-  });
-
-  test('throws if string is unterminated', () => {
-    expect(() => stringReader('"aaa', 0, context)).toThrow();
-  });
-
-  test('throws if string with escaped quote chars is unterminated', () => {
-    expect(() => stringReader('"aaa\\"', 0, context)).toThrow();
-  });
-});
 
 describe('jsonTokenizer', () => {
